+++ src/main/java/ru/vyarus/dropwizard/guice/module/context/info/DropwizardBundleItemInfo.java
@@ -1,0 +1,14 @@
+package ru.vyarus.dropwizard.guice.module.context.info;
+
+import ru.vyarus.dropwizard.guice.module.context.info.sign.DisableSupport;
+
+/**
+ * Dropwizard bundle configuration information.
+ * <p>
+ * Note that only directly registered dropwizard bundles (through guicey api) are tracked!
+ *
+ * @author Vyacheslav Rusakov
+ * @since 24.07.2019
+ */
+public interface DropwizardBundleItemInfo extends InstanceItemInfo, DisableSupport {
+}