+++ CHANGELOG.md
@@ -44,7 +44,10 @@
         as multiple instances of the same bundle class could be registered. For simplicity,
         ItemId equals method consider class-only id's equal to any type instance id.
     - Add bundle loops detection: as multiple bundle instances allowed loops are highly possible
-        Entire bundle chain is provided in exception to simplify fixing loops.                                             
+        Entire bundle chain is provided in exception to simplify fixing loops.
+* Add support for dropwizard bundles tracking: they could be registered in guice bundle (with `.dropwizardBundles()`)
+    and in guicey bundle. Registered bundles could be disabled in main bundle (with .disableDropwizardBundle or 
+    custom predicate) and checked for duplicates (same as guicey bundles and guice modules)                                                     
 
 
 Main breaking changes were caused by: